Add cart to user
Add cart property to user and default to empty array
Add cart update method to add product and qty in user controller
Update routes to route to new method

TODO: Need to find a way to see if product is already in cart. Perhaps
handle on the front end?
Verification for request so that empty cart update can't be sent (though
this should be handled on front end)

Next steps:
Virtuals for total items in cart/total price of all items
Remove old cart resource and related code